+++ JCasADi/src/main/java/de/dhbw/rahmlab/casadi/api/core/wrapper/bool/BooleanVector.java
@@ -163,7 +163,7 @@ public class BooleanVector {
      *
      * @param n the number of boolean values to reserve space for
      */
-    public void reserve(long n) {
+    public void reserve(int n) {
         this.stdVectorBool.reserve(n);
     }
